refactor(cli): extract pipeline + sync.status + helpers (#691) — v1.3.80
Closes #691 (#arch-h8 follow-up).

Second pass extracting business logic from cli.py. Builds on #611
(which moved synthesize_estimate_report + _adapter_status).

Moves:
- cmd_all → llmwiki/pipeline.py:run_pipeline (110-LOC pipeline runner)
- cmd_sync_status + _resolve_key_exists → llmwiki/sync/status.py
  (sync observability — state, quarantine, orphan detection)
- _load_schedule_config + _should_run_after_sync →
  llmwiki/config_schedule.py (config-policy, not CLI)
- _synthesize_list_pending + _synthesize_complete →
  llmwiki/synth/cli_helpers.py (consistent with synth/estimate.py)

cli.py shrinks 1,228 → 942 LOC (-286). All originals re-exported
under their underscored names so existing import paths keep working.

tests/test_cli_observability.py updated: 5 sites now also patch
llmwiki.sync.status.REPO_ROOT (the function moved out of cli.py
namespace).

Pure refactor. Full pytest suite passes.
