+++ llmwiki/__init__.py
@@ -15,7 +15,7 @@ Public API:
     - llmwiki.mcp.server.main()       — MCP server (stdio)
 """
 
-__version__ = "1.3.79"
+__version__ = "1.3.80"
 __author__ = "Pratiyush"
 __license__ = "MIT"
